Explain current solution for attribute extraction
